@@ -17,10 +17,11 @@ class Pillar {
     Pillar(double x, double y, double heading, double radius);
     Pillar(double x, double y, double radius);
     //~Pillar();
-    Pose2D getPose();
+    Pose2D& getPose();
     double getX();
     double getY();
     [[nodiscard]] double getRadius() const;
+    void setRadius(double newRadius);
 
     void setPosition(const Pose2D& newPosition);
 
